requirements.txt
numpy==1.21.2
pandas==1.3.3
streamlit==1.0.0
Pillow==8.4.0
